refactor: mail 폴더로 관리하도록 변경
